Split tee profiling harness tests by concern
Replace the monolithic tee profiling harness test module with focused
fixture, folded-summary, sink, driver, and worker test modules backed by a
shared redaction helper.

Regenerate snapshots under the new module names and remove the obsolete
monolithic snapshot file.
